rag_knowledge_base/requirements.txt
agentscope>=1.0.16
python-docx
PyPDF2
pandas
openpyxl